requirements.txt
tensorflow>=2.5.0
opencv-python>=4.5.0
numpy>=1.19.0
pillow>=8.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
flask>=2.0.0
requests>=2.26.0